解决生成代码界面的编辑时，没有修改表名的情况下点保存时，表名丢失的问题。并在后端增加一个判断：仅当表名发生了变化时，才更新字段设置。

+++ Admin.NET/Admin.NET.Core/Service/CodeGen/SysCodeGenService.cs
@@ -87,13 +87,18 @@ public class SysCodeGenService : IDynamicApiController, ITransient
         var isExist = await _db.Queryable<SysCodeGen>().AnyAsync(u => u.TableName == input.TableName && u.Id != input.Id);
         if (isExist) throw Oops.Oh(ErrorCodeEnum.D1400);
 
+        var oldRecord = await _db.Queryable<SysCodeGen>().FirstAsync(u => u.Id == input.Id);
+
         if (input.TableUniqueList?.Count > 0) input.TableUniqueConfig = JSON.Serialize(input.TableUniqueList);
         var codeGen = input.Adapt<SysCodeGen>();
         await _db.Updateable(codeGen).ExecuteCommandAsync();
 
-        // 更新配置表
-        await _codeGenConfigService.DeleteCodeGenConfig(codeGen.Id);
-        _codeGenConfigService.AddList(GetColumnList(input.Adapt<AddCodeGenInput>()), codeGen);
+        // 仅当数据表名称发生了变化，才更新配置表
+        if (oldRecord.TableName != input.TableName)
+        {
+            await _codeGenConfigService.DeleteCodeGenConfig(codeGen.Id);
+            _codeGenConfigService.AddList(GetColumnList(input.Adapt<AddCodeGenInput>()), codeGen);
+        }
     }
 
     /// <summary>
